NCAP/requirements.txt
digi-xbee==1.3.0
PyMySQL==0.9.3
pyserial==3.4
six==1.14.0
srp==1.0.15
